requirements.txt
pytesseract==0.3.13
Pillow>=11.0.0
PyPDF2==3.0.1
pdf2image==1.17.0
python-docx==1.1.2
mcp==1.1.2
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0

